requirements.txt
click
numpy
pyyaml
requests
rich
whiptail-dialogs
